requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
crispy-bootstrap4==2025.6
cryptography==45.0.7
dj-database-url==3.0.1
Django==5.2.6
django-allauth==65.11.1
django-crispy-forms==2.4
django-heroku==0.3.1
django-mysql==4.17.0
django-pwa==2.0.1
django-resized==1.0.3
ecdsa==0.19.1
forex-python==1.9.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jsonpickle==4.1.1
jwt==1.4.0
MarkupSafe==3.0.2
mysqlclient==2.2.7
packaging==25.0
paystack==1.5.0
paystack-api==0.1.2
paystackapi==2.1.3
pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.1.1
python-http-client==3.3.7
python_paystack==1.2.0
requests==2.32.5
sendgrid==6.12.4
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
validators==0.35.0
vercel==0.3.4
vercel-sandbox==0.0.3
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.9.0
xlwt==1.3.0
